LiPD CLI : Scaffolding
- Created the overall layout of the program
- Split Unzip and Creating Tmp directory into two distinct steps. Unzip
should not be responsible for creating a tmp folder, and we need to be
able to get a tmp folder without doing the rest of the Unzip steps also.
- Set DIRECTORY_PATH as a global variable in all scripts so it’s at the
top and easy to locate for users.
-LiPD_CLI class: The command line interface that manages and calls
actions on the objects.
- LiPD_Library class: Maintains all LiPD objects. Allows you to take
action on all, subsets, or single LiPD objects.
- LiPD class: Maintains all info about one specific LiPD file. Unpacks
file, retrieves data, saves changes, etc.