LiPD Lint, ExtractTimeSeries
LiPD Lint: More accurate name switches. Before it was switching based
on a list of names from the entire lint document. This didn’t work well
because there were some synonyms that overlapped and were corrected to
the wrong name. Now, it checks synonyms by section so that we get the
corrected name specific to that section.

ExtractTimeSeries: No longer uses the TS_Library or TS objects. It
instead returns a dictionary of data, that can be saved to a local
variable.