NOAA: Data Section: Variables
- Sometimes NOAA files are improperly formatted and the variables are
bunched together. (Supposed to be separated by spaces, but not always
true.) Parser would fail to match the variables in the Variables
section to the variables in the Data section. Instead of looking for an
exact match, this will check if the variables IS IN the line, which
seems to be a better way.