Skim CSV after import
The way that csv data was being skimmed into the LiPD object during
import no longer worked with all the recent changes. (At least not in a
way that was intuitive or modular) . This new process waits for csv
data to be imported into the metadata, and then skims and compiles it
into the LiPD object self. This is better than directly reading from
the csv files because it allows us to track the column variable names
with the value data. If we tried to do this another way, we’d have to
search the csv and metadata side by side which gets messy.