WIP: modified excel data sheet parser
Not fully tested. This parser will improve upon the only one. It
doesn’t rely as heavily on spreadsheet template items that may or may
not be present. At the very least, it depends on a Metadata section,
and a Numeric data section. Other pieces are optional. The function is
more universal and designed to work with model/summary/ensemble sheets
whether they are paleo or chron. Tested working with older files.
Currently testing with files that have updated formatting.