New excel parser & more
- Excel is now being parsed via a new method with checkpoint markers
(similar to the noaa parser) and is working much better. It’s not so
picky in metadata sections and can write to csv by rows so that no
outlying data is missed or ignored. This has been tested on an older
workbook, but needs more testing on a variety.
- LiPD v1.2: There was a mistake in the creation of how loadLipds()
updated older LiPD files to 1.2. I had thought that tables were only
given one measurement table, when they are actually able to have
multiple. I made changes to various areas of loadLipds() that now
reflect measurement tables as a list of items.
- There is a new regex that reads in excel worksheet names and creates
metadata for each worksheet. These worksheet names are now
standardized, and will be used to place the parsed excel data into the
LiPD structure.