GeoJSON FeatureCollection
New user files have examples data sets with multiple site locations in
them. This is reflected in GeoJSON as a FeatureCollection: a list where
each site has its own set. The parser now checks for the max number of
elevation, lat, or lon coordinates given, and assumes that the max is
the number of sites being parsed. Then parses each one accordingly.