Missing Values, and rogue cells
 - XLRD row call sometimes grabs more cells that hold nothing. This
causes a mismatch between the amount of variables that we have, and the
number of cells that were returned in the row. I now count our
variables, and slice the row for the amount of variables we have to
avoid a CSV with a bunch of blank columns.
- Missing values of -999 or other custom values were not matching and
being substituted correctly because the error handling was accounting
only for  strings.
- Parse geo location has been split into two functions. One function to
parse one location, and one function to parse multiple locations. It
was getting too hairy to try to do it all in one function and kept
parsing incorrectly for a single location.