Refactor import/export
There were a lot of processes that used two similar functions to do
that same thing for paleo and chron. This was because paleo and chron
used to be different. Paleo was less “open” than chron and held less
data. Now that they’re the same structures, I merged the functions to
work for either.