createDataframe() and cross-language naming
- createDataframe() was made so that users can create a data frame from
a nested array of ensemble data. Limited testing
- Cross-language functions names: In order to make it easier for users
trying to use LiPD utilities in assorted languages (R, Matlab, Python),
this is the first installation of changes to make all function calls
the same. Underlying language conventions are still going to persist
for functions that are behind-the-scenes that the user should not be
accessing, but any user-facing functions will follow camelCase standard
and be identical across languages.