Utilities: Multi-file loader
Changed the file loader. You can select 1 or more files in the file
chooser, or choose to load a full directory.

If you load a full directory, it’ll track and sort all files or
relevant file types (excel, lpd, txt) into the global space, so you can
run whichever function and have everything available.

This also allows files to be continually added and tracked in the
global space. Running the setDir() again should add files to the
workspace, instead or replacing the files currently in the workplace.