lpd_noaa changes
12.02.16 Meeting : Chris & Nick

Made some changes that were discussed in our last meeting.

- Original_Source_URL and Investigators are assumed to be in the root
of the metadata. However, if they are not found in the root, it will
look for a publication “dataCitation” and use values from there

- keyConversion() recursive fn has been removed. Certain keys had a 1:M
relationship between NOAA and LIPD, so context is important when
converting keys. I made another key conversion function that converts
based on section and key.

- 9 part variables: archiveType is now a global value and is used when
writing out the variables under “Archive”.  Also, since the 9 part
variables are comma separated, it now checks the strings being written
and replaces all commas with semi-colons to avoid confusion.

- Dataset DOI is not the same as Publication DOIs. These are not
treated completely different and separated. DOIs are no longer a global
variable.