Excel: Ensemble csv data fix
The ensemble tables were only writing out csv files with two columns
instead of the 1000+ that they usually have. For ensemble tables, it
counts the number of data columns based on the first for of data in the
data portion of the sheet. All other table types will still count the
columns based on the number of metadata entries + variables found.