Update to include NOAA's Requests
NOAA requested we add a number of small changes to the LiPD to NOAA
parser. I have added all the changes that are within reach. Any
requests not included are dependent on the data given in the LiPD file.
The converter will include as much data as possible when available.

- Tree Species Code, Tree Species Name: added to the key map.
- Most_recent_year, Earliest_year, Time_unit: added to the key map.
Time_unit will default to “AD” when not provided with “year” keys
given.
- Coordinates: swapped to correct positions.
- Variables Section: no longer uses fixed spacing for short_name.
- Authors & Investigators: checks for singular and plural keys (i.e.
Author and Authors) just to be safe. New function for parsing author
and investigator data. Handles more cases and makes sure data gets
parsed.
- Publication: Full_Citations are no longer duplicated. If two
publications contain the same Full_Citation, only the first
Full_Citation is written.
-Year variable: Year is not prioritized and written first in the
Variables section and the Data section.
- Online Resource: Includes a reference link for the NOAA txt file and
the LiPD file