Inferred Data, Validator Integration, and more!
- Max/min/median/mean and resolution are calculated and stored into
each PaleoData column (with exceptions) during the readLipds() process.
These data are calculated with numpy, BUT stored as python lists, due
to errors with not being able to write numpy data to file.

- (Not yet active!) Full validator integration into Python. Trigger
validation during readLipds(), and also available as a separate
function with validator() with 'detailed' or 'brief' results

- Missing Elevation in NOAA: Fixed a bug that used .pop() on the geo
coordinates list, and removed the elevation value permanently.

- TSid : During writeLipd(), checks all columns for a TSid key. It
generates a TSid for any column that does not have one.

- TSO: No longer excludes age, depth, year columns when creating TSOs.

- LiPD Archive Contents : LiPD objects now retain a list of all files
that were in the original LiPD file during the readLipd() process. This
is needed for when we call the validator API.

- Created a generate_tsid() that is (currently) used in writeLipd() and
excel() functions.

- Changed the parameters of zipper() to be my own naming declared
values. The documentation and requirements for the parameters to
shutil.make_archive() are ambiguous and don't make sense.

- NOAA: New "WDCPaleoURL" key to indicate where this LiPD file is being
stored in NOAA's archive. Use writeLipd() to write out newly modified
LiPD file.

-  NOAA: Added "notes" variable to the 9-part data section. This is not
a required field, but may help to differentiate near-identical columns.

- Some files had column numbers stored as strings. Cast values to int
before trying to use them.

- Removed "reload LiPD" part of the writeLipd() function. Now, instead
of saving the workspace to LiPD files and reloading those files back
into the workspace, it ONLY does the saving process.

- writeLipds() accepts a directory path parameter, if you'd like to
choose where to write your LiPD files.

- writeLipds() prints quantity of files being written, and which file
is currently being written.

- writeLipds() will only delete an old LiPD file if it  is in the
target directory that it is trying to write the new LiPD file. If
directories are different, then the original file is not deleted

- Split up get_src_or_dst() into multiple functions. Makes the code
easier to follow. Accommodates the changes to writeLipd().