Excel additions
- Added readLipd() and writeLipd() hooks to for each excel file. This
is because readLipd() and writeLipd() have all the support for
cleaning, updating, and ensuring a LiPD file is well-made.
- Added inferred data calculations to each file. Since each excel file
runs through readLipd() before the end of the process, it will also
gain the benefits of inferred data calculations for each paleoData
measurement column.
- Added more keys to the mapping from Excel to LiPD.
- Added mappings, regexes that will created nested dictionaries for
interpretation and calibration data when given.
- Added a global “verbose” variable in init.py to dictate when print
statements need to be ignored.