NOAA to LiPD is no longer a "separate" process
NOAA to LiPD has been integrated with the LiPD Library. Now, all data
that goes through the NOAA converter is passed from the LiPD Library,
converted, and passed back into the LiPD library. This means all the
data is consistent, up-to-date, cleaned, and is read and written in a
uniform way. (Whereas the NOAA converter previously had its own reading
and writing steps)