Activated validate(): lipd.net API live
- Routes for TSids and Validate through the lipd.net API are live and
working. Activated the validate() function in the python utilities so
files can be validated in batch.
- Since readLipds() does some manipulation and corrections to the JSON
metadata, I had to be specific about which metadata to send to the
validator, or else it doesn’t give the proper results back. At the end
of the read() in the LiPD object, I set the “validator friendly”
metadata to the object’s self.
- Since http calls the the validator may take a long time if a large
set of LiPD files are loaded, I removed the validate() call from
readLipds(). I don’t want to assume the user’s workflow and
automatically run the validator.
- Added a “fetch_new_data” parameter to validate(). This parameter will
default to “True”. The benefit is that if you run validate() on a large
set of files and it takes a long time, you don’t want to have to do
that again in the same session. The results of validate() are stored in
each LiPD object and retrieved whenever “fetch_new_results” is set to
False.
-  Changed some print statements to make things easier to read.