Ensemble handling
Ensemble tables were breaking the write_csv function because they were
not part of the test cases when the core functions were changed.

There are currently two cases for ensemble tables. 1. with depth column
2. without depth column

This fix accounts for both situations. However, case 2 is not yet
handled during readLipd() and will break. That fix is coming next.