Fix ensemble tables readLipds()
- Ensemble tables were not handled properly when trying to read in CSV
data to json columns. Wasn’t prepared to handle “1 column” ensembles
(without a depth column). Added new code to handle this case.
- Every table calls for inferred calculations. This was running for
ensemble tables which is bad! Ensembles are so large that it was
causing my computer to slow down to a halt, and then eventually exit
python with code 137 (out of memory). SO, I made sure to skip inferred
calculation when processing ensemble tables, as it is not needed
anyways.