Long list of NOAA requests, all complete where possible
1. The ordering of column headers of the NOAA text templates (defined
by the Variables) are not in sync with the actual Data columns in NOAA
text templates. Nam-GB.txt is a case in point. Please see Appendix A.
below for the full example.
 — BUG: fix whatever is wrong

2. Ordering of columns in the JSON-LD and NOAA text template need to be
the same.
—IGNORE. They rescinded this request

3. Data columns of the NOAA text templates are space-delimited, not
tab-delimited
— MY FAULT: Drop all the fixed spacing formats and just do
tab-delimited spacing instead.

4.Missing: TREE SPECIES CODES
— MY FAULT: I was pulling the info from the root. It’s not in the root!
pull sensorGenus and sensorSpecies from the column level. Make a list
of all unique 4-character codes and write out each.

5. Variables are ambiguous for 6 templates for which the
Orignal_Source_URL is not WDS NOAA Paleo or Pangaea
— NO SOLUTION: Not much we can do here. We’ve talked about this one.

6. Units for the LiPD "minYear" and "maxYear" tags - how can we
programmatically and consistently obtain the units from the LiPD JSON-LD

GIVE PREFERENCE TO AGE (YEARS BP)

if there is more than one year column or more than one age column, then
take the largest maximum and the lowest minimum and put that into the
NOAA file

— MY FAULT : Forget minYear and maxYear in the LiPD root. We don’t want
them. INSTEAD,

option 1 : look for a YEAR column in the data tables. calculate the
maxYear, minYear, timeUnit (units) and write that into the NOAA file

OR

option 2 : look for a AGE column in the data tables. calculate the
maxYear, minYear, timeUnit (units) and write that into the NOAA file

7.  For datasets with chronology information, this is included in the
NOAA text template as an extra "Data" section, rather than in the
chronology section.  Please place this information in the "Chronology"
section.
— MY FAULT: code is being repeated for PALEO and CHRON in the same way.
just change it so it writes the header differently for CHRON tables.

8. Remove “Filename:” tag in the Variables section it does not belong
there.
— MY FAULT  - completed

9. Online_Resource & Online_Resource_Description NOAA Txt tag not
correct
When there are two NOAA Txt files for a given .lpd archive (named
...-1.txt and ...-2.txt),
— MY FAULT - will fix

10.  615 of 778 NOAA text template files do not contain information for
tag Study_Name
standardize - generate study name using “ author - site name - year”

11. The Investigators tag is missing in some NOAA text templates. Only
165 of 778 NOAA text template files contain this tag
solution :
take all the publication authors from the first publication and put it
as the “investigators” field
use semi-colons to separate each person in the investigators field

12.LiPD QCNotes - It would be nice to get the LiPD QCNotes put into the
Description_Notes_and_Keywords Section.
— QUESTION  :  “QCNotes” is a LiPD variable at the column level. What
do we do when there are possibly more than one entry of these over many
columns?

solution:
aggregate all QCNotes, and put them together into one QCNote field.
Separate them by semi-colons

13. Publications in general: "Authors:", all investigators are
separated by the word "and", which we recommend be replaced by
semi-colons..
— MY FAULT: not that important, but I could write a function to replace
the “and”s with commas instead.
use semi-colons to separate each person in the investigators field

14. Nick Problem
source inconsistencies. nick is looking into it

15. Almost all of the d18O variables are missing their “Material”
component.
— QUESTION : up for discussion, mostly a nick issue

16. In all NOAA text templates, all the “floats” in the vars need to be
changed to “N”s.
— MY FAULT : fix it.

17. In NOAA text templates, there are a few cases where P.I.’s name is
repeated in Investigators tag.
See: Ocean2kHR-PacificPalmyracombinedCobb2003.txt
— NOT MY FAULT: This is how the LiPD file shows the data and not my
fault
solution - nick is looking into this

18. Nick Problem
solution - nick is going to fix this

19. . In NOAA text templates, in the definition of variables, Archive
is always missing.
— QUESITON : add archiveType to both root level and column level???
solution : provide a list of our archiveTypes
column level : provide whatever is found
root overall level : provide whatever is found, OR provide whatever is
the MAJORITY archiveType counted in the columns.

20. LiPD filenames -  order of author, year, site is not consistent
solution : standardize - generate study name using “ author - site name
- year”

21. The Keyword, Funding, and Chronology Sections are completely
missing from all NOAA text templates.  We request that these tags be
included even if there are no values.
— solution : sections are only created when there is data to fill them.
I can change that.

22. In LiPD JSON-LD, What does the “values” tag contain - this for our
own information  - what is the giant block of data? If it is in the
CSV, it is not clear why is it here... and in some cases the count of
values it contains does not match the count of rows in a CSV.
— OUR FAULT: this is the resolution values array, which we are going to
remove.