Issue #6: Python: MeasurementTable Number doesn't match between Excel and LiPD file
- The table indexes and location recorded in the table metadata did not
match the actual indexes that the data was being written out to in the
JSONLD hierarchy. It also differed from the CSV filename that its
corresponding data was being written to. The underlying problem is
caused by the unordered nature of Python dictionaries.

For example, a table at paleo1measurement1 may be read into memory at
paleo1measurement2, while still having a “paleoDataTableName” of
“paleo1measurement1”. Similarly, the CSV filename would be
“DatasetName.paleo1measurement2.csv” which does not match either.

The fix is Ordered Dictionaries. I added them throughout the readLipd
process to maintain the original structure and ensure that
paleo1measurement1 stays at paleo1measurement1, and so forth.