Validator API and Csv error msg
- The validator results functions were receiving back some empty files
somehow. When I tried to get the filename, it would break on a
KeyError. Now, I loop until I find a filename instead of relying on the
first index.
- CSVs have had a hard time with some mismatched column counts and
errors when comparing with the jsonld data. Added an error message that
is more targeted for this situation, and a general fall-through
exception after that.