Removed plural read/write functions
Instead of having a singular and plural version of all read and write
functions, it now determines whether to read a single file or directory
based on the path given. If no path is given, a prompt asks how many
file(s) / directory you want to read.

If you look on the write side, the function defaults to writing all
files in the workspace. If you pass a filename as a parameter, then it
will only write that one file.