Issue #12 : Decimal types - Benchmarks, coordinates
- Decimal types were being loaded in by the demjson module to load in
long precision floats. Demjson loads in the jsonld file as a python
dictionary, so we were getting Decimal types from the beginning of the
load process. However, demjson has a hook that lets you set the
“decode_float” as a float, instead of a Decimal.
- Benchmarks are currently calculated for major functions only. Ex:
read / write / validate / extractTs. These are logged in the main
debug.log file and also in a new benchmark.log file. Benchmark.log ONLY
has benchmark reports.
- Changed the RotatingFileHandler for loggers to not create any log
backups. Only use the one that is currently on the system, and
overwrite when the max file size is reached.