Optional path parameter for readLipd()
- Some more function documentation
- isDirectory() : an R workaround to determining if a given path is a
directory or a file. Nothing built-in to R.
- Added optional path parameter to readLipd() in case user wants to
point directly to the directory or file that they want to load
- Fixed bug: Reached an error when trying to read a path that did not
contain any LiPD files. The read functions would continue and go out of
bounds when it found out the array didn’t have any LiPD filenames to
read. Now, check that the LiPD file list is not empty before trying to
read in.