Bugs: filename conflicts, getting started in R
- The old file “TimeSeries.py” seems like it was not changed to
“timeseries.py” in the newest git commit and pypi push. This is causing
import errors because the module “does not exist” when it’s looking for
the wrong module name. I renamed the file to timeseries_fns.py, just to
give it a different name entirely.
- Started working on the regex for the filterTs and queryTs for R
- The “missing()” function wasn’t working right for determining if the
user passed in a path parameter to readLipd(), so I went a different
route with it. Working now.
- added libraries: tools, base, stringr to the dependencies NAMESPACE
- Added v1.3 to the list of valid lipdVersion values.
- Added some comments to extractTs in R. Also made it so that the depth
age and year columns are now extracted to their own time series entries
too.