Mirroring Python in R: pt 4
- Everything seems to be working with some test cases. Reads and writes
LiPD files and inner data just fine. Bagit working through the system
call. A couple of small things to fix  (expanded folder should be “bag”
and not the data set name for v1.3) and other small mishaps (the crumbs
for the CSV filenames is missing an index number and camel casing; Need
to add crumbs as “tableName” as well.)