R:  Error in file(out, “wt”) : cannot open the connection
Possible fix for this bug. I’ve read a lot of solutions that don’t
really help. I think the problem is related to creating and unlinking
temp directories during readLipd / writeLipd. I was unlinking the R
session directory, which is used for other things as well. We’ll see if
it works.