Python: fixed excel bug
- Bug that caused LiPD files created by excel conversion to be corrupt
files. This was because the zipper() function had changed and the
parameters in excel() were old values that no longer created a valid
archive. Added some more 1.3 updates to the excel conversion, and added
more exception handling to functions that surrounded this problem (i.e.
lipd_read, lipd_write, read_jsonld, and excel())