Python: version updater bug
The lipd version updater was only looking for a “LiPDVersion” key.
There are a handful of variations of this key and it wasn’t catching
the other ones. This caused v1.3 files to be considered 1.0 files by
default (when it couldnt find the key) and then try to update a valid
file into an invalid file.