R: Wiki files don't have values in columns readLipd()
Wiki files store the “number” for each column as a string. All of the
utilities are expecting this to be a numeric or an integer. The values
where not being placed properly in the columns because using the string
caused NULL data to come back.

The fix was to check for a character “number” value, and cast to
numeric as needed.