R: loose key check for filterTs and queryTs
Currently, if you were looking to filter a time series by
“interpretation_variable == P/T”, then you would have to also know
which interpretation index to look at “interpretation1_variable == P/T”
which is very specific. Now, you can do a loose check on the key, and
then all the matches are added to the output. So “interpretation ==
P/T” will catch “interpretation1_variable == P/T” or
“interpretation2_variable == P/T”, etc