Python: Decimal precision, file size, wiki bug
- Fixed bug caused by files that come from the wiki. They use “number”
fields in columns as a string data type. These are supposed to be
integers. Changed this during CSV file write as it was causing errors.
- Fixed decimal precision during CSV write. Function accepts a row that
is about to be written, and does some workaround magic for integers,
floats, and floats that use scientific notation. Rounds everything to 3
decimals. Deborah was creating models, and using the new addModel()
function to add them to datasets. However, models that have values with
20+ decimal precision were creating files that were 10x larger.