Python: addTable
A supplement to addModel. addTable will let you add a single summary or
measurement table (ensemble and distribution are a bit more
complicated) by choosing the associated saved CSV file, and entering
metadata via prompts. It will determine the next available locations to
store the table in the dataset, and ask you to choose one. It returns
the original dataset with the table added.