Python: Few updates
- LiPD to NOAA: This process now requires a project name and version
number via input prompts. This is used to create the WDS paleo URL
- addTable : Completed for now. It does all the essential steps to
adding in a data table via csv files.
- Jsons: Ran into some more json decode bugs. After some guess work, I
found out that the problematic files were encoded with latin-1
settings, and using latin-1 to decode was all that was needed. So,
reading jsonld files now attempts utf8 first, and then tries latin-1 if
that fails.