Py: extractTs(), global space, Fix tkinter bug
- Added new options to extractTs. You can choose meas, summ, ens, or
all tables to put into the output time series.
- Added global “raw” data from the extractTs. Now, instead of storing a
copy of the dataset in each object, we can just store one instance of
it in the global space. It’s still available for collapseTs later.
Caveat: There can only be one “run” of the time series data in the
global space at one time.
- The Tkinter bug is fixed!!! No more rogue python processes. Since
this is a deep down python bug that will likely never be fixed, it
affected all gui modules out there. I had to come up with my own
workaround to the issue. I stored the file browse and directory browse
gui in their own scripts, and call on that script as a subprocess. I
catch the stdout byte string and normalize it. The subprocess will
close itself when ready so no more rogue processes!