Py: Timeseries support for summ and ens tables
- Create multiple timeseries’ and retain the raw data in the global
space
- Extract any table type you want, or all table types. Method data
attached for model tables. Each object contains the time series
metadata.
mode (paleoData/chronData)
tableType = “meas”, “summ”, “ens”
time_id  = 4.964527 (an arbitrary float number to differentiate
timeseries’ in global space in the instance of multiple extractions)
paleoNumber / chronNumber
modelNumber - when necessary
tableNumber

-CollapseTs now supports all tables as well. The initial “raw” dataset
is used from the global space. The structure is kept, but the data is
overwritten. This is because creating the dataset skeletons is a hairy
process.