Python: lpd_noaa updates pt.2
- Bug: Appended filenames for files with multiple data tables were not
outputting correctly. The “WDSPaleoUrl” entries that were placed in the
LiPD file metadata was only showing the regular filenames, and not an
appended filename.
- Added more parameters as URLs are evolving. There are now two
different URLs. The NOAA template url, and the LiPD dataset url. These
have the potential of ending up on different addresses on the NOAA FTP
server.
- Recursive function to remove all WDSPaleoUrl data from LiPD metadata
- Function to remove “originalDataURL” data from LiPD metadata root