R: collapseTs bug fixes for Force=True
Fixed some issues preventing time series data from being collapsed when attempting to use force=TRUE when tmp global data is not available.
It was calling functions that aren't needed when using force=TRUE. For example, it doesn't make sense to remove old tables from the raw data if you're not using raw data and those tables don't exist. Also, time series metadata is filled in with default values if it's not present. We make assumptions that it is paleoData measurement table in the first position.